# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaFileSummary": {
        "description": "One file a listing can put in `media:` as a file ref: a\nshared one under ``common-media/``, or one of the listing's own. A bare\nfile uploaded as-is, rather than a rendered mockup.",
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "kind": {
            "enum": [
              "image",
              "video"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "name",
          "file",
          "ref",
          "kind"
        ],
        "title": "MediaFileSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "etsy-listings",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/common-media/{name}/file": {
      "get": {
        "description": "The shared file at its own size, for the preview pane and the\nlightbox -- the bytes as they sit on disk, which are exactly what Etsy\nwould receive.",
        "operationId": "common_media_file_api_common_media__name__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Common Media File",
        "tags": [
          "media-files"
        ]
      }
    },
    "/api/common-media/{name}/thumbnail": {
      "get": {
        "operationId": "common_media_thumbnail_api_common_media__name__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Common Media Thumbnail",
        "tags": [
          "media-files"
        ]
      }
    },
...
    "/api/listing-templates/{template}/media-files": {
      "get": {
        "description": "The *This template* group (UI doc, *Existing components the template\neditor needs*): the listing template's own files, each with the ``./``\nref that names it from ``template.yaml``. Scoped to ``assets/``.",
        "operationId": "list_listing_template_media_files_api_listing_templates__template__media_files_get",
        "parameters": [
          {
            "in": "path",
            "name": "template",
            "required": true,
            "schema": {
              "title": "Template",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaFileSummary"
                  },
                  "title": "Response List Listing Template Media Files Api Listing Templates  Template  Media Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Listing Template Media Files",
        "tags": [
          "media-files"
        ]
      }
    },
    "/api/listing-templates/{template}/media-files/{path}/file": {
      "get": {
        "operationId": "listing_template_media_file_api_listing_templates__template__media_files__path__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "template",
            "required": true,
            "schema": {
              "title": "Template",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing Template Media File",
        "tags": [
          "media-files"
        ]
      }
    },
    "/api/listing-templates/{template}/media-files/{path}/thumbnail": {
      "get": {
        "description": "A listing template's own file (its ``./`` root, ADR-0047) -- the pictures on\nits card. The same boundary as a listing's: the directory also holds\n``template.yaml``.",
        "operationId": "listing_template_media_thumbnail_api_listing_templates__template__media_files__path__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "template",
            "required": true,
            "schema": {
              "title": "Template",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing Template Media Thumbnail",
        "tags": [
          "media-files"
        ]
      }
    },
...
    "/api/listings/{listing}/media-files": {
      "get": {
        "description": "The *This listing* group: the listing's own files, each with the\n``./`` ref that names it. A ``404`` for a listing that does not\nexist, rather than an empty group that would look like one with nothing\nin it.",
        "operationId": "list_listing_media_files_api_listings__listing__media_files_get",
        "parameters": [
          {
            "in": "path",
            "name": "listing",
            "required": true,
            "schema": {
              "title": "Listing",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaFileSummary"
                  },
                  "title": "Response List Listing Media Files Api Listings  Listing  Media Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Listing Media Files",
        "tags": [
          "media-files"
        ]
      }
    },
    "/api/listings/{listing}/media-files/{path}/file": {
      "get": {
        "operationId": "listing_media_file_api_listings__listing__media_files__path__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "listing",
            "required": true,
            "schema": {
              "title": "Listing",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing Media File",
        "tags": [
          "media-files"
        ]
      }
    },
    "/api/listings/{listing}/media-files/{path}/thumbnail": {
      "get": {
        "operationId": "listing_media_thumbnail_api_listings__listing__media_files__path__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "listing",
            "required": true,
            "schema": {
              "title": "Listing",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing Media Thumbnail",
        "tags": [
          "media-files"
        ]
      }
    },
...
  }
}